requirements.txt
Django==5.2.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
pillow==10.4.0 # for image processing
django-filter==25.1
drf-nested-routers==0.94.2
drf-spectacular==0.28.0